requirements.txt
requests
playwright
scikit-learn
nltk
pdfminer.six
pandas
python-dotenv
